Remove DeviceSetFactory, instead pass a DeviceSet around
We now lazily instantiate the DeviceSet, so object creation is
free and error-less. This means we can always create one and pass
around. This is nice because it allows us to pass on the DeviceSet
to other places that need it (e.g. Graph) without needing to have
a full Runtime object.