Add DeviceSet to DaemonConfig and fill it in
For the daemon and for the main test runtime we just create plain
device-mapper DeviceSets, but for newTestRuntime() we return a
wrapper on the main test DeviceSet that uses a prefix. This way
each test doesn't have to create new loopback mounts and whatnot.

We also make sure to properly shut down the DeviceSet at the end of
the test run to make sure we're not leaving anything behind.