Replace AUFS with device mapper for the copy-on-write container FS
This uses the DeviceSet to replace the previous use of AUFS mounts
for the container FS. We create an devmapper snapshot for every
image based on the parent images snapshot, then we create snapshots
for the container based on the image.

To commit and to generate diffs from the image we now need to
traverse the full container fs to look for differences.